Update VSync if VSync hint changes
Currently, if an application wants to toggle VSync, they'd have to tear
down the renderer and recreate it. This patch fixes that by adding hint
callbacks to all renderers if SDL_HINT_RENDER_VSYNC changes.

This is the same as the patch in #3673, except it applies to all
renderers (including PSP, even thought it seems that the VSync flag is
disabled for that renderer). Furthermore, the renderer flags also change
as well, which #3673 didn't do.

Closes #3673.